Module aws_sdk_config::error

Modules

conformance_pack_template_validation_exception

See ConformancePackTemplateValidationException

insufficient_delivery_policy_exception

See InsufficientDeliveryPolicyException

insufficient_permissions_exception

See InsufficientPermissionsException

invalid_configuration_recorder_name_exception

See InvalidConfigurationRecorderNameException

invalid_delivery_channel_name_exception

See InvalidDeliveryChannelNameException

invalid_expression_exception

See InvalidExpressionException

invalid_limit_exception

See InvalidLimitException

invalid_next_token_exception

See InvalidNextTokenException

invalid_parameter_value_exception

See InvalidParameterValueException

invalid_recording_group_exception

See InvalidRecordingGroupException

invalid_result_token_exception

See InvalidResultTokenException

invalid_role_exception

See InvalidRoleException

invalid_s3_key_prefix_exception

See InvalidS3KeyPrefixException

invalid_s3_kms_key_arn_exception

See InvalidS3KmsKeyArnException

invalid_sns_topic_arn_exception

See InvalidSnsTopicArnException

invalid_time_range_exception

See InvalidTimeRangeException

last_delivery_channel_delete_failed_exception

See LastDeliveryChannelDeleteFailedException

limit_exceeded_exception

See LimitExceededException

max_active_resources_exceeded_exception

See MaxActiveResourcesExceededException

max_number_of_config_rules_exceeded_exception

See MaxNumberOfConfigRulesExceededException

max_number_of_configuration_recorders_exceeded_exception

See MaxNumberOfConfigurationRecordersExceededException

max_number_of_conformance_packs_exceeded_exception

See MaxNumberOfConformancePacksExceededException

max_number_of_delivery_channels_exceeded_exception

See MaxNumberOfDeliveryChannelsExceededException

max_number_of_organization_config_rules_exceeded_exception

See MaxNumberOfOrganizationConfigRulesExceededException

max_number_of_organization_conformance_packs_exceeded_exception

See MaxNumberOfOrganizationConformancePacksExceededException

max_number_of_retention_configurations_exceeded_exception

See MaxNumberOfRetentionConfigurationsExceededException

no_available_configuration_recorder_exception

See NoAvailableConfigurationRecorderException

no_available_delivery_channel_exception

See NoAvailableDeliveryChannelException

no_available_organization_exception

See NoAvailableOrganizationException

no_running_configuration_recorder_exception

See NoRunningConfigurationRecorderException

no_such_bucket_exception

See NoSuchBucketException

no_such_config_rule_exception

See NoSuchConfigRuleException

no_such_config_rule_in_conformance_pack_exception

See NoSuchConfigRuleInConformancePackException

no_such_configuration_aggregator_exception

See NoSuchConfigurationAggregatorException

no_such_configuration_recorder_exception

See NoSuchConfigurationRecorderException

no_such_conformance_pack_exception

See NoSuchConformancePackException

no_such_delivery_channel_exception

See NoSuchDeliveryChannelException

no_such_organization_config_rule_exception

See NoSuchOrganizationConfigRuleException

no_such_organization_conformance_pack_exception

See NoSuchOrganizationConformancePackException

no_such_remediation_configuration_exception

See NoSuchRemediationConfigurationException

no_such_remediation_exception_exception

See NoSuchRemediationExceptionException

no_such_retention_configuration_exception

See NoSuchRetentionConfigurationException

organization_access_denied_exception

See OrganizationAccessDeniedException

organization_all_features_not_enabled_exception

See OrganizationAllFeaturesNotEnabledException

organization_conformance_pack_template_validation_exception

See OrganizationConformancePackTemplateValidationException

oversized_configuration_item_exception

See OversizedConfigurationItemException

remediation_in_progress_exception

See RemediationInProgressException

resource_concurrent_modification_exception

See ResourceConcurrentModificationException

resource_in_use_exception

See ResourceInUseException

resource_not_discovered_exception

See ResourceNotDiscoveredException

resource_not_found_exception

See ResourceNotFoundException

too_many_tags_exception

See TooManyTagsException

validation_exception

See ValidationException

Structs

BatchGetAggregateResourceConfigError

BatchGetResourceConfigError

ConformancePackTemplateValidationException

You have specified a template that is not valid or supported.

DeleteAggregationAuthorizationError

DeleteConfigRuleError

DeleteConfigurationAggregatorError

DeleteConfigurationRecorderError

DeleteConformancePackError

DeleteDeliveryChannelError

DeleteEvaluationResultsError

DeleteOrganizationConfigRuleError

DeleteOrganizationConformancePackError

DeletePendingAggregationRequestError

DeleteRemediationConfigurationError

DeleteRemediationExceptionsError

DeleteResourceConfigError

DeleteRetentionConfigurationError

DeleteStoredQueryError

DeliverConfigSnapshotError

DescribeAggregateComplianceByConfigRulesError

DescribeAggregateComplianceByConformancePacksError

DescribeAggregationAuthorizationsError

DescribeComplianceByConfigRuleError

DescribeComplianceByResourceError

DescribeConfigRuleEvaluationStatusError

DescribeConfigRulesError

DescribeConfigurationAggregatorSourcesStatusError

DescribeConfigurationAggregatorsError

DescribeConfigurationRecorderStatusError

DescribeConfigurationRecordersError

DescribeConformancePackComplianceError

DescribeConformancePackStatusError

DescribeConformancePacksError

DescribeDeliveryChannelStatusError

DescribeDeliveryChannelsError

DescribeOrganizationConfigRuleStatusesError

DescribeOrganizationConfigRulesError

DescribeOrganizationConformancePackStatusesError

DescribeOrganizationConformancePacksError

DescribePendingAggregationRequestsError

DescribeRemediationConfigurationsError

DescribeRemediationExceptionsError

DescribeRemediationExecutionStatusError

DescribeRetentionConfigurationsError

GetAggregateComplianceDetailsByConfigRuleError

GetAggregateConfigRuleComplianceSummaryError

GetAggregateConformancePackComplianceSummaryError

GetAggregateDiscoveredResourceCountsError

GetAggregateResourceConfigError

GetComplianceDetailsByConfigRuleError

GetComplianceDetailsByResourceError

GetComplianceSummaryByConfigRuleError

GetComplianceSummaryByResourceTypeError

GetConformancePackComplianceDetailsError

GetConformancePackComplianceSummaryError

GetDiscoveredResourceCountsError

GetOrganizationConfigRuleDetailedStatusError

GetOrganizationConformancePackDetailedStatusError

GetResourceConfigHistoryError

GetStoredQueryError

InsufficientDeliveryPolicyException

Your Amazon S3 bucket policy does not permit AWS Config to

InsufficientPermissionsException

Indicates one of the following errors:

InvalidConfigurationRecorderNameException

You have provided a configuration recorder name that is not

InvalidDeliveryChannelNameException

The specified delivery channel name is not valid.

InvalidExpressionException

The syntax of the query is incorrect.

InvalidLimitException

The specified limit is outside the allowable range.

InvalidNextTokenException

The specified next token is invalid. Specify the

InvalidParameterValueException

One or more of the specified parameters are invalid. Verify

InvalidRecordingGroupException

AWS Config throws an exception if the recording group does not contain a valid list of resource types. Invalid values might also be incorrectly formatted.

InvalidResultTokenException

The specified ResultToken is invalid.

InvalidRoleException

You have provided a null or empty role ARN.

InvalidS3KeyPrefixException

The specified Amazon S3 key prefix is not valid.

InvalidS3KmsKeyArnException

The specified Amazon KMS Key ARN is not valid.

InvalidSnsTopicArnException

The specified Amazon SNS topic does not exist.

InvalidTimeRangeException

The specified time range is not valid. The earlier time is not

LastDeliveryChannelDeleteFailedException

You cannot delete the delivery channel you specified because

LimitExceededException

For StartConfigRulesEvaluation API, this exception

ListAggregateDiscoveredResourcesError

ListDiscoveredResourcesError

ListStoredQueriesError

ListTagsForResourceError

MaxActiveResourcesExceededException

You have reached the limit (100,000) of active custom resource types in your account.

MaxNumberOfConfigRulesExceededException

Failed to add the AWS Config rule because the account already

MaxNumberOfConfigurationRecordersExceededException

You have reached the limit of the number of recorders you can

MaxNumberOfConformancePacksExceededException

You have reached the limit (6) of the number of conformance packs in an account (6 conformance pack with 25 AWS Config rules per pack).

MaxNumberOfDeliveryChannelsExceededException

You have reached the limit of the number of delivery channels

MaxNumberOfOrganizationConfigRulesExceededException

You have reached the limit of the number of organization config rules you can create.

MaxNumberOfOrganizationConformancePacksExceededException

You have reached the limit (6) of the number of organization conformance packs in an account (6 conformance pack with 25 AWS Config rules per pack per account).

MaxNumberOfRetentionConfigurationsExceededException

Failed to add the retention configuration because a retention configuration with that name already exists.

NoAvailableConfigurationRecorderException

There are no configuration recorders available to provide the

NoAvailableDeliveryChannelException

There is no delivery channel available to record

NoAvailableOrganizationException

Organization is no longer available.

NoRunningConfigurationRecorderException

There is no configuration recorder running.

NoSuchBucketException

The specified Amazon S3 bucket does not exist.

NoSuchConfigRuleException

One or more AWS Config rules in the request are invalid. Verify

NoSuchConfigRuleInConformancePackException

AWS Config rule that you passed in the filter does not exist.

NoSuchConfigurationAggregatorException

You have specified a configuration aggregator that does not exist.

NoSuchConfigurationRecorderException

You have specified a configuration recorder that does not

NoSuchConformancePackException

You specified one or more conformance packs that do not exist.

NoSuchDeliveryChannelException

You have specified a delivery channel that does not

NoSuchOrganizationConfigRuleException

You specified one or more organization config rules that do not exist.

NoSuchOrganizationConformancePackException

AWS Config organization conformance pack that you passed in the filter does not exist.

NoSuchRemediationConfigurationException

You specified an AWS Config rule without a remediation configuration.

NoSuchRemediationExceptionException

You tried to delete a remediation exception that does not exist.

NoSuchRetentionConfigurationException

You have specified a retention configuration that does not exist.

OrganizationAccessDeniedException

For PutConfigurationAggregator API, you can see this exception for the following reasons:

OrganizationAllFeaturesNotEnabledException

AWS Config resource cannot be created because your organization does not have all features enabled.

OrganizationConformancePackTemplateValidationException

You have specified a template that is not valid or supported.

OversizedConfigurationItemException

The configuration item size is outside the allowable range.

PutAggregationAuthorizationError

PutConfigRuleError

PutConfigurationAggregatorError

PutConfigurationRecorderError

PutConformancePackError

PutDeliveryChannelError

PutEvaluationsError

PutExternalEvaluationError

PutOrganizationConfigRuleError

PutOrganizationConformancePackError

PutRemediationConfigurationsError

PutRemediationExceptionsError

PutResourceConfigError

PutRetentionConfigurationError

PutStoredQueryError

RemediationInProgressException

Remediation action is in progress. You can either cancel execution in AWS Systems Manager or wait and try again later.

ResourceConcurrentModificationException

Two users are trying to modify the same query at the same time. Wait for a moment and try again.

ResourceInUseException

You see this exception in the following cases:

ResourceNotDiscoveredException

You have specified a resource that is either unknown or has not

ResourceNotFoundException

You have specified a resource that does not exist.

SelectAggregateResourceConfigError

SelectResourceConfigError

StartConfigRulesEvaluationError

StartConfigurationRecorderError

StartRemediationExecutionError

StopConfigurationRecorderError

TagResourceError

TooManyTagsException

You have reached the limit of the number of tags you can use. You have more than 50 tags.

UntagResourceError

ValidationException

The requested action is not valid.

Enums

BatchGetAggregateResourceConfigErrorKind

BatchGetResourceConfigErrorKind

DeleteAggregationAuthorizationErrorKind

DeleteConfigRuleErrorKind

DeleteConfigurationAggregatorErrorKind

DeleteConfigurationRecorderErrorKind

DeleteConformancePackErrorKind

DeleteDeliveryChannelErrorKind

DeleteEvaluationResultsErrorKind

DeleteOrganizationConfigRuleErrorKind

DeleteOrganizationConformancePackErrorKind

DeletePendingAggregationRequestErrorKind

DeleteRemediationConfigurationErrorKind

DeleteRemediationExceptionsErrorKind

DeleteResourceConfigErrorKind

DeleteRetentionConfigurationErrorKind

DeleteStoredQueryErrorKind

DeliverConfigSnapshotErrorKind

DescribeAggregateComplianceByConfigRulesErrorKind

DescribeAggregateComplianceByConformancePacksErrorKind

DescribeAggregationAuthorizationsErrorKind

DescribeComplianceByConfigRuleErrorKind

DescribeComplianceByResourceErrorKind

DescribeConfigRuleEvaluationStatusErrorKind

DescribeConfigRulesErrorKind

DescribeConfigurationAggregatorSourcesStatusErrorKind

DescribeConfigurationAggregatorsErrorKind

DescribeConfigurationRecorderStatusErrorKind

DescribeConfigurationRecordersErrorKind

DescribeConformancePackComplianceErrorKind

DescribeConformancePackStatusErrorKind

DescribeConformancePacksErrorKind

DescribeDeliveryChannelStatusErrorKind

DescribeDeliveryChannelsErrorKind

DescribeOrganizationConfigRuleStatusesErrorKind

DescribeOrganizationConfigRulesErrorKind

DescribeOrganizationConformancePackStatusesErrorKind

DescribeOrganizationConformancePacksErrorKind

DescribePendingAggregationRequestsErrorKind

DescribeRemediationConfigurationsErrorKind

DescribeRemediationExceptionsErrorKind

DescribeRemediationExecutionStatusErrorKind

DescribeRetentionConfigurationsErrorKind

GetAggregateComplianceDetailsByConfigRuleErrorKind

GetAggregateConfigRuleComplianceSummaryErrorKind

GetAggregateConformancePackComplianceSummaryErrorKind

GetAggregateDiscoveredResourceCountsErrorKind

GetAggregateResourceConfigErrorKind

GetComplianceDetailsByConfigRuleErrorKind

GetComplianceDetailsByResourceErrorKind

GetComplianceSummaryByConfigRuleErrorKind

GetComplianceSummaryByResourceTypeErrorKind

GetConformancePackComplianceDetailsErrorKind

GetConformancePackComplianceSummaryErrorKind

GetDiscoveredResourceCountsErrorKind

GetOrganizationConfigRuleDetailedStatusErrorKind

GetOrganizationConformancePackDetailedStatusErrorKind

GetResourceConfigHistoryErrorKind

GetStoredQueryErrorKind

ListAggregateDiscoveredResourcesErrorKind

ListDiscoveredResourcesErrorKind

ListStoredQueriesErrorKind

ListTagsForResourceErrorKind

PutAggregationAuthorizationErrorKind

PutConfigRuleErrorKind

PutConfigurationAggregatorErrorKind

PutConfigurationRecorderErrorKind

PutConformancePackErrorKind

PutDeliveryChannelErrorKind

PutEvaluationsErrorKind

PutExternalEvaluationErrorKind

PutOrganizationConfigRuleErrorKind

PutOrganizationConformancePackErrorKind

PutRemediationConfigurationsErrorKind

PutRemediationExceptionsErrorKind

PutResourceConfigErrorKind

PutRetentionConfigurationErrorKind

PutStoredQueryErrorKind

SelectAggregateResourceConfigErrorKind

SelectResourceConfigErrorKind

StartConfigRulesEvaluationErrorKind

StartConfigurationRecorderErrorKind

StartRemediationExecutionErrorKind

StopConfigurationRecorderErrorKind

TagResourceErrorKind

UntagResourceErrorKind