Enum aws_sdk_config::error::PutOrganizationConformancePackErrorKind

#[non_exhaustive]
pub enum PutOrganizationConformancePackErrorKind {
    InsufficientPermissionsException(InsufficientPermissionsException),
    MaxNumberOfOrganizationConformancePacksExceededException(MaxNumberOfOrganizationConformancePacksExceededException),
    NoAvailableOrganizationException(NoAvailableOrganizationException),
    OrganizationAccessDeniedException(OrganizationAccessDeniedException),
    OrganizationAllFeaturesNotEnabledException(OrganizationAllFeaturesNotEnabledException),
    OrganizationConformancePackTemplateValidationException(OrganizationConformancePackTemplateValidationException),
    ResourceInUseException(ResourceInUseException),
    ValidationException(ValidationException),
    Unhandled(Box<dyn Error + Send + Sync + 'static>),
}

Types of errors that can occur for the PutOrganizationConformancePack operation.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

InsufficientPermissionsException(InsufficientPermissionsException)

Indicates one of the following errors:

For PutConfigRule, the rule cannot be created because the IAM role assigned to Config lacks permissions to perform the config:Put* action.

For PutConfigRule, the Lambda function cannot be invoked. Check the function ARN, and check the function's permissions.

For PutOrganizationConfigRule, organization config rule cannot be created because you do not have permissions to call IAM GetRole action or create a service linked role.

For PutConformancePack and PutOrganizationConformancePack, a conformance pack cannot be created because you do not have permissions:

To call IAM GetRole action or create a service linked role.

To read Amazon S3 bucket.

Tuple Fields of InsufficientPermissionsException

0: InsufficientPermissionsException

MaxNumberOfOrganizationConformancePacksExceededException(MaxNumberOfOrganizationConformancePacksExceededException)

You have reached the limit (6) of the number of organization conformance packs in an account (6 conformance pack with 25 Config rules per pack per account).

Tuple Fields of MaxNumberOfOrganizationConformancePacksExceededException

0: MaxNumberOfOrganizationConformancePacksExceededException

NoAvailableOrganizationException(NoAvailableOrganizationException)

Organization is no longer available.

Tuple Fields of NoAvailableOrganizationException

0: NoAvailableOrganizationException

OrganizationAccessDeniedException(OrganizationAccessDeniedException)

For PutConfigurationAggregator API, you can see this exception for the following reasons:

No permission to call EnableAWSServiceAccess API

The configuration aggregator cannot be updated because your Amazon Web Services Organization management account or the delegated administrator role changed. Delete this aggregator and create a new one with the current Amazon Web Services Organization.

The configuration aggregator is associated with a previous Amazon Web Services Organization and Config cannot aggregate data with current Amazon Web Services Organization. Delete this aggregator and create a new one with the current Amazon Web Services Organization.

You are not a registered delegated administrator for Config with permissions to call ListDelegatedAdministrators API. Ensure that the management account registers delagated administrator for Config service principle name before the delegated administrator creates an aggregator.

For all OrganizationConfigRule and OrganizationConformancePack APIs, Config throws an exception if APIs are called from member accounts. All APIs must be called from organization master account.

Tuple Fields of OrganizationAccessDeniedException

0: OrganizationAccessDeniedException

OrganizationAllFeaturesNotEnabledException(OrganizationAllFeaturesNotEnabledException)

Config resource cannot be created because your organization does not have all features enabled.

Tuple Fields of OrganizationAllFeaturesNotEnabledException

0: OrganizationAllFeaturesNotEnabledException

OrganizationConformancePackTemplateValidationException(OrganizationConformancePackTemplateValidationException)

You have specified a template that is not valid or supported.

Tuple Fields of OrganizationConformancePackTemplateValidationException

0: OrganizationConformancePackTemplateValidationException

ResourceInUseException(ResourceInUseException)

You see this exception in the following cases:

For DeleteConfigRule, Config is deleting this rule. Try your request again later.

For DeleteConfigRule, the rule is deleting your evaluation results. Try your request again later.

For DeleteConfigRule, a remediation action is associated with the rule and Config cannot delete this rule. Delete the remediation action associated with the rule before deleting the rule and try your request again later.

For PutConfigOrganizationRule, organization config rule deletion is in progress. Try your request again later.

For DeleteOrganizationConfigRule, organization config rule creation is in progress. Try your request again later.

For PutConformancePack and PutOrganizationConformancePack, a conformance pack creation, update, and deletion is in progress. Try your request again later.

For DeleteConformancePack, a conformance pack creation, update, and deletion is in progress. Try your request again later.

Tuple Fields of ResourceInUseException

0: ResourceInUseException

ValidationException(ValidationException)

The requested action is not valid.

For PutStoredQuery, you will see this exception if there are missing required fields or if the input value fails the validation, or if you are trying to create more than 300 queries.

For GetStoredQuery, ListStoredQuery, and DeleteStoredQuery you will see this exception if there are missing required fields or if the input value fails the validation.

Tuple Fields of ValidationException

0: ValidationException

Unhandled(Box<dyn Error + Send + Sync + 'static>)

An unexpected error, e.g. invalid JSON returned by the service or an unknown error code

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static>