Module aws_sdk_config::model

Modules

account_aggregation_source

See AccountAggregationSource

aggregate_compliance_by_config_rule

See AggregateComplianceByConfigRule

aggregate_compliance_by_conformance_pack

See AggregateComplianceByConformancePack

aggregate_compliance_count

See AggregateComplianceCount

aggregate_conformance_pack_compliance

See AggregateConformancePackCompliance

aggregate_conformance_pack_compliance_count

See AggregateConformancePackComplianceCount

aggregate_conformance_pack_compliance_filters

See AggregateConformancePackComplianceFilters

aggregate_conformance_pack_compliance_summary

See AggregateConformancePackComplianceSummary

aggregate_conformance_pack_compliance_summary_filters

See AggregateConformancePackComplianceSummaryFilters

aggregate_evaluation_result

See AggregateEvaluationResult

aggregate_resource_identifier

See AggregateResourceIdentifier

aggregated_source_status

See AggregatedSourceStatus

aggregation_authorization

See AggregationAuthorization

base_configuration_item

See BaseConfigurationItem

compliance

See Compliance

compliance_by_config_rule

See ComplianceByConfigRule

compliance_by_resource

See ComplianceByResource

compliance_contributor_count

See ComplianceContributorCount

compliance_summary

See ComplianceSummary

compliance_summary_by_resource_type

See ComplianceSummaryByResourceType

config_export_delivery_info

See ConfigExportDeliveryInfo

config_rule

See ConfigRule

config_rule_compliance_filters

See ConfigRuleComplianceFilters

config_rule_compliance_summary_filters

See ConfigRuleComplianceSummaryFilters

config_rule_evaluation_status

See ConfigRuleEvaluationStatus

config_snapshot_delivery_properties

See ConfigSnapshotDeliveryProperties

config_stream_delivery_info

See ConfigStreamDeliveryInfo

configuration_aggregator

See ConfigurationAggregator

configuration_item

See ConfigurationItem

configuration_recorder

See ConfigurationRecorder

configuration_recorder_status

See ConfigurationRecorderStatus

conformance_pack_compliance_filters

See ConformancePackComplianceFilters

conformance_pack_compliance_summary

See ConformancePackComplianceSummary

conformance_pack_detail

See ConformancePackDetail

conformance_pack_evaluation_filters

See ConformancePackEvaluationFilters

conformance_pack_evaluation_result

See ConformancePackEvaluationResult

conformance_pack_input_parameter

See ConformancePackInputParameter

conformance_pack_rule_compliance

See ConformancePackRuleCompliance

conformance_pack_status_detail

See ConformancePackStatusDetail

delivery_channel

See DeliveryChannel

delivery_channel_status

See DeliveryChannelStatus

evaluation

See Evaluation

evaluation_result

See EvaluationResult

evaluation_result_identifier

See EvaluationResultIdentifier

evaluation_result_qualifier

See EvaluationResultQualifier

execution_controls

See ExecutionControls

external_evaluation

See ExternalEvaluation

failed_delete_remediation_exceptions_batch

See FailedDeleteRemediationExceptionsBatch

failed_remediation_batch

See FailedRemediationBatch

failed_remediation_exception_batch

See FailedRemediationExceptionBatch

field_info

See FieldInfo

grouped_resource_count

See GroupedResourceCount

member_account_status

See MemberAccountStatus

organization_aggregation_source

See OrganizationAggregationSource

organization_config_rule

See OrganizationConfigRule

organization_config_rule_status

See OrganizationConfigRuleStatus

organization_conformance_pack

See OrganizationConformancePack

organization_conformance_pack_detailed_status

See OrganizationConformancePackDetailedStatus

organization_conformance_pack_status

See OrganizationConformancePackStatus

organization_custom_rule_metadata

See OrganizationCustomRuleMetadata

organization_managed_rule_metadata

See OrganizationManagedRuleMetadata

organization_resource_detailed_status_filters

See OrganizationResourceDetailedStatusFilters

pending_aggregation_request

See PendingAggregationRequest

query_info

See QueryInfo

recording_group

See RecordingGroup

relationship

See Relationship

remediation_configuration

See RemediationConfiguration

remediation_exception

See RemediationException

remediation_exception_resource_key

See RemediationExceptionResourceKey

remediation_execution_status

See RemediationExecutionStatus

remediation_execution_step

See RemediationExecutionStep

remediation_parameter_value

See RemediationParameterValue

resource_count

See ResourceCount

resource_count_filters

See ResourceCountFilters

resource_filters

See ResourceFilters

resource_identifier

See ResourceIdentifier

resource_key

See ResourceKey

resource_value

See ResourceValue

retention_configuration

See RetentionConfiguration

scope

See Scope

source

See Source

source_detail

See SourceDetail

ssm_controls

See SsmControls

static_value

See StaticValue

status_detail_filters

See StatusDetailFilters

stored_query

See StoredQuery

stored_query_metadata

See StoredQueryMetadata

tag

See Tag

Structs

AccountAggregationSource

A collection of accounts and regions.

AggregateComplianceByConfigRule

Indicates whether an AWS Config rule is compliant based on

AggregateComplianceByConformancePack

Provides aggregate compliance of the conformance pack. Indicates whether a conformance pack is compliant based on the name of the conformance pack, account ID, and region.

AggregateComplianceCount

Returns the number of compliant and noncompliant rules for one

AggregateConformancePackCompliance

Provides the number of compliant and noncompliant rules within a conformance pack.

AggregateConformancePackComplianceCount

The number of conformance packs that are compliant and noncompliant.

AggregateConformancePackComplianceFilters

Filters the conformance packs based on an account ID, region, compliance type, and the name of the conformance pack.

AggregateConformancePackComplianceSummary

Provides a summary of compliance based on either account ID or region.

AggregateConformancePackComplianceSummaryFilters

Filters the results based on account ID and region.

AggregateEvaluationResult

The details of an AWS Config evaluation for an account ID and

AggregateResourceIdentifier

The details that identify a resource that is collected by AWS Config aggregator, including the resource type, ID, (if available) the custom resource name, the source account, and source region.

AggregatedSourceStatus

The current sync status between the source and the aggregator

AggregationAuthorization

An object that represents the authorizations granted to

BaseConfigurationItem

The detailed configuration of a specified resource.

Compliance

Indicates whether an AWS resource or AWS Config rule is

ComplianceByConfigRule

Indicates whether an AWS Config rule is compliant. A rule is

ComplianceByResource

Indicates whether an AWS resource that is evaluated according

ComplianceContributorCount

The number of AWS resources or AWS Config rules responsible for

ComplianceSummary

The number of AWS Config rules or AWS resources that are

ComplianceSummaryByResourceType

The number of AWS resources of a specific type that are

ConfigExportDeliveryInfo

Provides status of the delivery of the snapshot or the

ConfigRule

An AWS Config rule represents an AWS Lambda function that you

ConfigRuleComplianceFilters

Filters the compliance results based on account ID, region,

ConfigRuleComplianceSummaryFilters

Filters the results based on the account IDs and

ConfigRuleEvaluationStatus

Status information for your AWS managed Config rules. The

ConfigSnapshotDeliveryProperties

Provides options for how often AWS Config delivers

ConfigStreamDeliveryInfo

A list that contains the status of the delivery of the

ConfigurationAggregator

The details about the configuration aggregator, including

ConfigurationItem

A list that contains detailed configurations of a specified

ConfigurationRecorder

An object that represents the recording of configuration

ConfigurationRecorderStatus

The current status of the configuration recorder.

ConformancePackComplianceFilters

Filters the conformance pack by compliance types and AWS Config rule names.

ConformancePackComplianceSummary

Summary includes the name and status of the conformance pack.

ConformancePackDetail

Returns details of a conformance pack. A conformance pack is a collection of AWS Config rules and remediation actions that can be easily deployed in an account and a region.

ConformancePackEvaluationFilters

Filters a conformance pack by AWS Config rule names, compliance types, AWS resource types, and resource IDs.

ConformancePackEvaluationResult

The details of a conformance pack evaluation. Provides AWS Config rule and AWS resource type that was evaluated, the compliance of the conformance pack, related time stamps, and supplementary information.

ConformancePackInputParameter

Input parameters in the form of key-value pairs for the conformance pack, both of which you define.

ConformancePackRuleCompliance

Compliance information of one or more AWS Config rules within a conformance pack. You can filter using AWS Config rule names and compliance types.

ConformancePackStatusDetail

Status details of a conformance pack.

DeliveryChannel

The channel through which AWS Config delivers notifications and

DeliveryChannelStatus

The status of a specified delivery channel.

Evaluation

Identifies an AWS resource and indicates whether it complies

EvaluationResult

The details of an AWS Config evaluation. Provides the AWS

EvaluationResultIdentifier

Uniquely identifies an evaluation result.

EvaluationResultQualifier

Identifies an AWS Config rule that evaluated an AWS resource,

ExecutionControls

The controls that AWS Config uses for executing remediations.

ExternalEvaluation

Identifies an AWS resource and indicates whether it complies with the AWS Config rule that it was evaluated against.

FailedDeleteRemediationExceptionsBatch

List of each of the failed delete remediation exceptions with specific reasons.

FailedRemediationBatch

List of each of the failed remediations with specific reasons.

FailedRemediationExceptionBatch

List of each of the failed remediation exceptions with specific reasons.

FieldInfo

Details about the fields such as name of the field.

GroupedResourceCount

The count of resources that are grouped by the group name.

MemberAccountStatus

Organization config rule creation or deletion status in each member account. This includes the name of the rule, the status, error code and error message when the rule creation or deletion failed.

OrganizationAggregationSource

This object contains regions to set up the aggregator and an IAM

OrganizationConfigRule

An organization config rule that has information about config rules that AWS Config creates in member accounts.

OrganizationConfigRuleStatus

Returns the status for an organization config rule in an organization.

OrganizationConformancePack

An organization conformance pack that has information about conformance packs that AWS Config creates in member accounts.

OrganizationConformancePackDetailedStatus

Organization conformance pack creation or deletion status in each member account.

OrganizationConformancePackStatus

Returns the status for an organization conformance pack in an organization.

OrganizationCustomRuleMetadata

An object that specifies organization custom rule metadata such as resource type, resource ID of AWS resource, Lamdba function ARN,

OrganizationManagedRuleMetadata

An object that specifies organization managed rule metadata such as resource type and ID of AWS resource along with the rule identifier.

OrganizationResourceDetailedStatusFilters

Status filter object to filter results based on specific member account ID or status type for an organization conformance pack.

PendingAggregationRequest

An object that represents the account ID and region of an

QueryInfo

Details about the query.

RecordingGroup

Specifies the types of AWS resource for which AWS Config

Relationship

The relationship of the related resource to the main

RemediationConfiguration

An object that represents the details about the remediation configuration that includes the remediation action, parameters, and data to execute the action.

RemediationException

An object that represents the details about the remediation exception. The details include the rule name, an explanation of an exception, the time when the exception will be deleted, the resource ID, and resource type.

RemediationExceptionResourceKey

The details that identify a resource within AWS Config, including the resource type and resource ID.

RemediationExecutionStatus

Provides details of the current status of the invoked remediation action for that resource.

RemediationExecutionStep

Name of the step from the SSM document.

RemediationParameterValue

The value is either a dynamic (resource) value or a static value. You must select either a dynamic value or a static value.

ResourceCount

An object that contains the resource type and the number of

ResourceCountFilters

Filters the resource count based on account ID, region, and resource type.

ResourceFilters

Filters the results by resource account ID, region, resource ID, and resource name.

ResourceIdentifier

The details that identify a resource that is discovered by AWS

ResourceKey

The details that identify a resource within AWS Config, including

ResourceValue

The dynamic value of the resource.

RetentionConfiguration

An object with the name of the retention configuration and the retention period in days. The object stores the configuration for data retention in AWS Config.

Scope

Defines which resources trigger an evaluation for an AWS Config

Source

Provides the AWS Config rule owner (AWS or customer), the rule

SourceDetail

Provides the source and the message types that trigger AWS

SsmControls

AWS Systems Manager (SSM) specific remediation controls.

StaticValue

The static value of the resource.

StatusDetailFilters

Status filter object to filter results based on specific member account ID or status type for an organization config rule.

StoredQuery

Provides the details of a stored query.

StoredQueryMetadata

Returns details of a specific query.

Tag

The tags for the resource. The metadata that you apply to a resource to help you categorize and organize them.

Enums

AggregateConformancePackComplianceSummaryGroupKey

AggregatedSourceStatusType

AggregatedSourceType

ChronologicalOrder

ComplianceType

ConfigRuleComplianceSummaryGroupKey

ConfigRuleState

ConfigurationItemStatus

ConformancePackComplianceType

ConformancePackState

DeliveryStatus

EventSource

MaximumExecutionFrequency

MemberAccountRuleStatus

MessageType

OrganizationConfigRuleTriggerType

OrganizationResourceDetailedStatus

OrganizationResourceStatus

OrganizationRuleStatus

Owner

RecorderStatus

RemediationExecutionState

RemediationExecutionStepState

RemediationTargetType

ResourceCountGroupKey

ResourceType

ResourceValueType