Struct aws_sdk_config::model::ConfigurationItem

#[non_exhaustive]
pub struct ConfigurationItem {

    pub version: Option<String>,
    pub account_id: Option<String>,
    pub configuration_item_capture_time: Option<Instant>,
    pub configuration_item_status: Option<ConfigurationItemStatus>,
    pub configuration_state_id: Option<String>,
    pub configuration_item_md5_hash: Option<String>,
    pub arn: Option<String>,
    pub resource_type: Option<ResourceType>,
    pub resource_id: Option<String>,
    pub resource_name: Option<String>,
    pub aws_region: Option<String>,
    pub availability_zone: Option<String>,
    pub resource_creation_time: Option<Instant>,
    pub tags: Option<HashMap<String, String>>,
    pub related_events: Option<Vec<String>>,
    pub relationships: Option<Vec<Relationship>>,
    pub configuration: Option<String>,
    pub supplementary_configuration: Option<HashMap<String, String>>,

}

A list that contains detailed configurations of a specified resource.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

version: Option<String>

The version number of the resource configuration.

account_id: Option<String>

The 12-digit AWS account ID associated with the resource.

configuration_item_capture_time: Option<Instant>

The time when the configuration recording was initiated.

configuration_item_status: Option<ConfigurationItemStatus>

The configuration item status. The valid values are:

OK – The resource configuration has been updated

ResourceDiscovered – The resource was newly discovered

ResourceNotRecorded – The resource was discovered but its configuration was not recorded since the recorder excludes the recording of resources of this type

ResourceDeleted – The resource was deleted

ResourceDeletedNotRecorded – The resource was deleted but its configuration was not recorded since the recorder excludes the recording of resources of this type

The CIs do not incur any cost.

configuration_state_id: Option<String>

An identifier that indicates the ordering of the configuration items of a resource.

configuration_item_md5_hash: Option<String>

Unique MD5 hash that represents the configuration item's state.

You can use MD5 hash to compare the states of two or more configuration items that are associated with the same resource.

arn: Option<String>

Amazon Resource Name (ARN) associated with the resource.

resource_type: Option<ResourceType>

The type of AWS resource.

resource_id: Option<String>

The ID of the resource (for example, sg-xxxxxx).

resource_name: Option<String>

The custom name of the resource, if available.

aws_region: Option<String>

The region where the resource resides.

availability_zone: Option<String>

The Availability Zone associated with the resource.

resource_creation_time: Option<Instant>

The time stamp when the resource was created.

tags: Option<HashMap<String, String>>

A mapping of key value tags associated with the resource.

related_events: Option<Vec<String>>

A list of CloudTrail event IDs.

A populated field indicates that the current configuration was initiated by the events recorded in the CloudTrail log. For more information about CloudTrail, see What Is AWS CloudTrail.

An empty field indicates that the current configuration was not initiated by any event. As of Version 1.3, the relatedEvents field is empty. You can access the LookupEvents API in the AWS CloudTrail API Reference to retrieve the events for the resource.

relationships: Option<Vec<Relationship>>

A list of related AWS resources.

configuration: Option<String>

The description of the resource configuration.

supplementary_configuration: Option<HashMap<String, String>>

Configuration attributes that AWS Config returns for certain resource types to supplement the information returned for the configuration parameter.

Implementations

impl ConfigurationItem

pub fn builder() -> Builder

Creates a new builder-style object to manufacture ConfigurationItem