Module aws_sdk_config::operation

Structs

BatchGetAggregateResourceConfig

Returns the current configuration items for resources that are present in your AWS Config aggregator. The operation also returns a list of resources that are not processed in the current request.

BatchGetResourceConfig

Returns the current configuration for one or more requested

DeleteAggregationAuthorization

Deletes the authorization granted to the specified

DeleteConfigRule

Deletes the specified AWS Config rule and all of its evaluation

DeleteConfigurationAggregator

Deletes the specified configuration aggregator and the

DeleteConfigurationRecorder

Deletes the configuration recorder.

DeleteConformancePack

Deletes the specified conformance pack and all the AWS Config rules, remediation actions, and all evaluation results within that

DeleteDeliveryChannel

Deletes the delivery channel.

DeleteEvaluationResults

Deletes the evaluation results for the specified AWS Config

DeleteOrganizationConfigRule

Deletes the specified organization config rule and all of its evaluation results from all member accounts in that organization.

DeleteOrganizationConformancePack

Deletes the specified organization conformance pack and all of the config rules and remediation actions from

DeletePendingAggregationRequest

Deletes pending authorization requests for a specified

DeleteRemediationConfiguration

Deletes the remediation configuration.

DeleteRemediationExceptions

Deletes one or more remediation exceptions mentioned in the resource keys.

DeleteResourceConfig

Records the configuration state for a custom resource that has been deleted. This API records a new ConfigurationItem with a ResourceDeleted status. You can retrieve the ConfigurationItems recorded for this resource in your AWS Config History.

DeleteRetentionConfiguration

Deletes the retention configuration.

DeleteStoredQuery

Deletes the stored query for a single AWS account and a single AWS Region.

DeliverConfigSnapshot

Schedules delivery of a configuration snapshot to the Amazon S3

DescribeAggregateComplianceByConfigRules

Returns a list of compliant and noncompliant rules with the

DescribeAggregateComplianceByConformancePacks

Returns a list of the conformance packs and their associated compliance status with the count of compliant and noncompliant AWS Config rules within each conformance pack.

DescribeAggregationAuthorizations

Returns a list of authorizations granted to various aggregator

DescribeComplianceByConfigRule

Indicates whether the specified AWS Config rules are compliant.

DescribeComplianceByResource

Indicates whether the specified AWS resources are compliant. If

DescribeConfigRuleEvaluationStatus

Returns status information for each of your AWS managed Config

DescribeConfigRules

Returns details about your AWS Config rules.

DescribeConfigurationAggregatorSourcesStatus

Returns status information for sources within an aggregator.

DescribeConfigurationAggregators

Returns the details of one or more configuration aggregators.

DescribeConfigurationRecorderStatus

Returns the current status of the specified configuration

DescribeConfigurationRecorders

Returns the details for the specified configuration recorders.

DescribeConformancePackCompliance

Returns compliance details for each rule in that conformance pack.

DescribeConformancePackStatus

Provides one or more conformance packs deployment status.

DescribeConformancePacks

Returns a list of one or more conformance packs.

DescribeDeliveryChannelStatus

Returns the current status of the specified delivery channel.

DescribeDeliveryChannels

Returns details about the specified delivery channel. If a

DescribeOrganizationConfigRuleStatuses

Provides organization config rule deployment status for an organization.

DescribeOrganizationConfigRules

Returns a list of organization config rules.

DescribeOrganizationConformancePackStatuses

Provides organization conformance pack deployment status for an organization.

DescribeOrganizationConformancePacks

Returns a list of organization conformance packs.

DescribePendingAggregationRequests

Returns a list of all pending aggregation requests.

DescribeRemediationConfigurations

Returns the details of one or more remediation configurations.

DescribeRemediationExceptions

Returns the details of one or more remediation exceptions. A detailed view of a remediation exception for a set of resources that includes an explanation of an exception and the time when the exception will be deleted.

DescribeRemediationExecutionStatus

Provides a detailed view of a Remediation Execution for a set of resources including state, timestamps for when steps for the remediation execution occur, and any error messages for steps that have failed.

DescribeRetentionConfigurations

Returns the details of one or more retention configurations. If

GetAggregateComplianceDetailsByConfigRule

Returns the evaluation results for the specified AWS Config

GetAggregateConfigRuleComplianceSummary

Returns the number of compliant and noncompliant rules for one

GetAggregateConformancePackComplianceSummary

Returns the count of compliant and noncompliant conformance packs across all AWS Accounts and AWS Regions in an aggregator. You can filter based on AWS Account ID or AWS Region.

GetAggregateDiscoveredResourceCounts

Returns the resource counts across accounts and regions that are present in your AWS Config aggregator. You can request the resource counts by providing filters and GroupByKey.

GetAggregateResourceConfig

Returns configuration item that is aggregated for your specific resource in a specific source account and region.

GetComplianceDetailsByConfigRule

Returns the evaluation results for the specified AWS Config

GetComplianceDetailsByResource

Returns the evaluation results for the specified AWS resource.

GetComplianceSummaryByConfigRule

Returns the number of AWS Config rules that are compliant and

GetComplianceSummaryByResourceType

Returns the number of resources that are compliant and the

GetConformancePackComplianceDetails

Returns compliance details of a conformance pack for all AWS resources that are monitered by conformance pack.

GetConformancePackComplianceSummary

Returns compliance details for the conformance pack based on the cumulative compliance results of all the rules in that conformance pack.

GetDiscoveredResourceCounts

Returns the resource types, the number of each resource type,

GetOrganizationConfigRuleDetailedStatus

Returns detailed status for each member account within an organization for a given organization config rule.

GetOrganizationConformancePackDetailedStatus

Returns detailed status for each member account within an organization for a given organization conformance pack.

GetResourceConfigHistory

Returns a list of configuration items for the specified

GetStoredQuery

Returns the details of a specific stored query.

ListAggregateDiscoveredResources

Accepts a resource type and returns a list of resource identifiers that are aggregated for a specific resource type across accounts and regions.

ListDiscoveredResources

Accepts a resource type and returns a list of resource

ListStoredQueries

Lists the stored queries for a single AWS account and a single AWS Region. The default is 100.

ListTagsForResource

List the tags for AWS Config resource.

PutAggregationAuthorization

Authorizes the aggregator account and region to collect data

PutConfigRule

Adds or updates an AWS Config rule for evaluating whether your

PutConfigurationAggregator

Creates and updates the configuration aggregator with the

PutConfigurationRecorder

Creates a new configuration recorder to record the selected

PutConformancePack

Creates or updates a conformance pack. A conformance pack is a collection of AWS Config rules that can be easily deployed in an account and a region and across AWS Organization.

PutDeliveryChannel

Creates a delivery channel object to deliver configuration

PutEvaluations

Used by an AWS Lambda function to deliver evaluation results to

PutExternalEvaluation

Add or updates the evaluations for process checks.

PutOrganizationConfigRule

Adds or updates organization config rule for your entire organization evaluating whether your AWS resources comply with your

PutOrganizationConformancePack

Deploys conformance packs across member accounts in an AWS Organization.

PutRemediationConfigurations

Adds or updates the remediation configuration with a specific AWS Config rule with the

PutRemediationExceptions

A remediation exception is when a specific resource is no longer considered for auto-remediation.

PutResourceConfig

Records the configuration state for the resource provided in the request.

PutRetentionConfiguration

Creates and updates the retention configuration with details

PutStoredQuery

Saves a new query or updates an existing saved query. The QueryName must be unique for a single AWS account and a single AWS Region.

SelectAggregateResourceConfig

Accepts a structured query language (SQL) SELECT command and an aggregator to query configuration state of AWS resources across multiple accounts and regions,

SelectResourceConfig

Accepts a structured query language (SQL) SELECT command, performs the corresponding search, and returns resource configurations matching the properties.

StartConfigRulesEvaluation

Runs an on-demand evaluation for the specified AWS Config rules

StartConfigurationRecorder

Starts recording configurations of the AWS resources you have

StartRemediationExecution

Runs an on-demand remediation for the specified AWS Config rules against the last known remediation configuration. It runs an execution against the current state of your resources. Remediation execution is asynchronous.

StopConfigurationRecorder

Stops recording configurations of the AWS resources you have selected to record in your AWS account.

TagResource

Associates the specified tags to a resource with the specified resourceArn. If existing tags on a resource are not specified in the request parameters, they are not changed.

UntagResource

Deletes specified tags from a resource.